PYGUARD/requirements.txt
altair @ file:///work/perseverance-python-buildout/croot/altair_1698866618831/work
attrs @ file:///work/perseverance-python-buildout/croot/attrs_1698845858439/work
blinker @ file:///work/perseverance-python-buildout/croot/blinker_1698846408677/work
Bottleneck @ file:///croot/bottleneck_1709069899917/work
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools @ file:///croot/cachetools_1713977085444/work
certifi @ file:///croot/certifi_1720453481653/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///work/perseverance-python-buildout/croot/click_1698845879718/work
dnspython==2.6.1
gitdb @ file:///tmp/build/80754af9/gitdb_1617117951232/work
GitPython @ file:///croot/gitpython_1720455034730/work
idna @ file:///croot/idna_1714398848350/work
Jinja2 @ file:///croot/jinja2_1716993405101/work
jsonschema @ file:///work/perseverance-python-buildout/croot/jsonschema_1701731705290/work
jsonschema-specifications @ file:///work/perseverance-python-buildout/croot/jsonschema-specifications_1701731664072/work
markdown-it-py @ file:///work/perseverance-python-buildout/croot/markdown-it-py_1698846045803/work
MarkupSafe @ file:///work/perseverance-python-buildout/croot/markupsafe_1707342994941/work
mdurl @ file:///work/perseverance-python-buildout/croot/mdurl_1698845653285/work
mkl-fft @ file:///work/perseverance-python-buildout/croot/mkl_fft_1698845673361/work
mkl-random @ file:///work/perseverance-python-buildout/croot/mkl_random_1698845720894/work
mkl-service==2.4.0
numexpr @ file:///work/perseverance-python-buildout/croot/numexpr_1698871031164/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp312-cp312-linux_x86_64.whl#sha256=1d700f51d8b4fa684d858c9e3b56b1656bc5c82b6b79ff08d4e3b491c430059f
packaging @ file:///croot/packaging_1710807400464/work
pandas @ file:///croot/pandas_1718308974269/work/dist/pandas-2.2.2-cp312-cp312-linux_x86_64.whl#sha256=92c518f7e09edd50b5caa5862636c51d6a29391803f3ada62f68aa52f27d8f92
pillow @ file:///croot/pillow_1721059439630/work
protobuf==4.25.3
psycopg2-binary==2.9.9
pyarrow @ file:///croot/pyarrow_1721664224170/work/python
pydeck @ file:///croot/pydeck_1706194064552/work
Pygments @ file:///work/perseverance-python-buildout/croot/pygments_1698846270603/work
pymongo==4.8.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.0.1
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///work/perseverance-python-buildout/croot/pyyaml_1698849903511/work
referencing @ file:///work/perseverance-python-buildout/croot/referencing_1701731622327/work
requests @ file:///croot/requests_1721410876868/work
rich @ file:///croot/rich_1720637495510/work
rpds-py @ file:///work/perseverance-python-buildout/croot/rpds-py_1698982686240/work
setuptools==72.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap @ file:///tmp/build/80754af9/smmap_1611694433573/work
streamlit==1.37.1
tenacity @ file:///croot/tenacity_1721222508356/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
toolz @ file:///work/perseverance-python-buildout/croot/toolz_1698846367152/work
tornado @ file:///croot/tornado_1718740109488/work
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///croot/urllib3_1718912636303/work
watchdog @ file:///croot/watchdog_1717166512516/work
wheel==0.43.0
